+++ src/renderer/components/layout/PlayBar/MiniWidthProgress.vue
@@ -1,52 +1,50 @@
 <template>
   <div :class="$style.player">
-    <div :class="$style.picContent" :aria-label="$t('player__pic_tip')" @contextmenu="handleToMusicLocation" @click="showPlayerDetail">
-      <img v-if="musicInfo.pic" :src="musicInfo.pic" decoding="async" @error="imgError">
-      <div v-else :class="$style.emptyPic">L<span>X</span></div>
+    <div :class="$style.progress">
+      <common-progress-bar v-if="!isShowPlayerDetail" edge :progress="progress" :handle-transition-end="handleTransitionEnd" :is-active-transition="isActiveTransition" />
     </div>
-    <div :class="$style.infoContent">
-      <div :class="$style.titleRow">
-        <div :class="$style.title" :aria-label="title + $t('copy_tip')" @click="handleCopy(title)">
-          {{ title }}
-        </div>
-        <button :class="[$style.commentBtn, {[$style.commentActive]: isShowPlayComment}]" :aria-label="$t('comment__show')" :title="$t('comment__show')" ignore-tip @click="toggleVisibleComment">
-          <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" width="100%" viewBox="0 0 24 24" space="preserve">
-            <use xlink:href="#icon-comment" />
-          </svg>
-        </button>
+    <div :class="$style.left">
+      <div :class="$style.picContent" :aria-label="$t('player__pic_tip')" @contextmenu="handleToMusicLocation" @click="showPlayerDetail">
+        <img v-if="musicInfo.pic" :class="{[$style.rotating]: isPlay}" :src="musicInfo.pic" decoding="async" @error="imgError">
+        <div v-else :class="$style.emptyPic">L<span>X</span></div>
       </div>
-      <div :class="$style.status">{{ singer || statusText }}</div>
-    </div>
-    <div :class="$style.centerZone">
-      <div :class="$style.transportRow">
-        <div :class="$style.playBtn" :aria-label="$t('player__prev')" @click="playPrev()">
-          <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" height="100%" viewBox="0 0 1024 1024" space="preserve">
-            <use xlink:href="#icon-prevMusic" />
-          </svg>
+      <div :class="$style.infoContent">
+        <div :class="$style.titleRow">
+          <div :class="$style.title" :aria-label="title + $t('copy_tip')" @click="handleCopy(title)">
+            {{ title }}
+          </div>
+          <button :class="[$style.commentBtn, {[$style.commentActive]: isShowPlayComment}]" :aria-label="$t('comment__show')" :title="$t('comment__show')" ignore-tip @click="toggleVisibleComment">
+            <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" width="100%" viewBox="0 0 24 24" space="preserve">
+              <use xlink:href="#icon-comment" />
+            </svg>
+          </button>
         </div>
-        <div :class="[$style.playBtn, $style.mainBtn]" :aria-label="isPlay ? $t('player__pause') : $t('player__play')" @click="togglePlay">
-          <svg v-if="isPlay" version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" height="100%" viewBox="0 0 1024 1024" space="preserve">
-            <use xlink:href="#icon-pause" />
-          </svg>
-          <svg v-else version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" height="100%" viewBox="0 0 1024 1024" space="preserve">
-            <use xlink:href="#icon-play" />
-          </svg>
-        </div>
-        <div :class="$style.playBtn" :aria-label="$t('player__next')" @click="playNext()">
-          <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" height="100%" viewBox="0 0 1024 1024" space="preserve">
-            <use xlink:href="#icon-nextMusic" />
-          </svg>
-        </div>
-      </div>
-      <div :class="$style.progressRow">
-        <span :class="$style.time">{{ nowPlayTimeStr }}</span>
-        <div :class="$style.progress">
-          <common-progress-bar v-if="!isShowPlayerDetail" :class-name="$style.progressBar" :progress="progress" :handle-transition-end="handleTransitionEnd" :is-active-transition="isActiveTransition" />
-        </div>
-        <span :class="$style.time">{{ maxPlayTimeStr }}</span>
+        <div :class="$style.status">{{ singer || statusText }}</div>
       </div>
     </div>
-    <control-btns />
+    <div :class="$style.playBtnContent">
+      <div :class="$style.playBtn" :aria-label="$t('player__prev')" @click="playPrev()">
+        <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" height="100%" viewBox="0 0 1024 1024" space="preserve">
+          <use xlink:href="#icon-prevMusic" />
+        </svg>
+      </div>
+      <div :class="[$style.playBtn, $style.mainBtn]" :aria-label="isPlay ? $t('player__pause') : $t('player__play')" @click="togglePlay">
+        <svg v-if="isPlay" version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" height="100%" viewBox="0 0 1024 1024" space="preserve">
+          <use xlink:href="#icon-pause" />
+        </svg>
+        <svg v-else version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" height="100%" viewBox="0 0 1024 1024" space="preserve">
+          <use xlink:href="#icon-play" />
+        </svg>
+      </div>
+      <div :class="$style.playBtn" :aria-label="$t('player__next')" @click="playNext()">
+        <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xlink="http://www.w3.org/1999/xlink" height="100%" viewBox="0 0 1024 1024" space="preserve">
+          <use xlink:href="#icon-nextMusic" />
+        </svg>
+      </div>
+    </div>
+    <div :class="$style.right">
+      <control-btns />
+    </div>
   </div>
 </template>
 
@@ -168,14 +166,13 @@ export default {
 .player {
   position: relative;
   height: @height-player;
-  border-top: 1px solid var(--color-200);
   box-shadow: 0 -10px 28px rgba(0, 0, 0, .06);
   box-sizing: border-box;
   display: flex;
   flex-flow: row nowrap;
   align-items: center;
-  contain: layout style;
-  padding: 8px 10px;
+  overflow: visible;
+  padding: 10px 10px 8px;
   z-index: 2;
   * {
     box-sizing: border-box;
@@ -191,6 +188,25 @@ export default {
     opacity: .96;
     z-index: -1;
   }
+}
+.progress {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 14px;
+  z-index: 3;
+}
+.left,
+.right {
+  flex: 1 1 0;
+  min-width: 0;
+  display: flex;
+  flex-flow: row nowrap;
+  align-items: center;
+}
+.right {
+  justify-content: flex-end;
 }
 
 .picContent {
@@ -219,16 +235,24 @@ export default {
   img {
     max-width: 100%;
     max-height: 100%;
-    transition: @transition-normal;
-    transition-property: border-color, transform, box-shadow;
-    // 圆形碟片样式（网易云风格）
     border-radius: 50%;
     border: 3px solid var(--color-950);
     box-shadow: 0 4px 14px rgba(0, 0, 0, 0.22);
+    animation: play-bar-vinyl-rotate 18s linear infinite;
+    animation-play-state: paused;
   }
 
-  &:hover img {
-    transform: scale(1.03);
+  .rotating {
+    animation-play-state: running;
+  }
+
+  @keyframes play-bar-vinyl-rotate {
+    from {
+      transform: rotate(0deg);
+    }
+    to {
+      transform: rotate(360deg);
+    }
   }
 
   .emptyPic {
@@ -257,7 +281,7 @@ export default {
 }
 
 .infoContent {
-  flex: 0 1 230px;
+  flex: 1 1 auto;
   min-width: 0;
   padding: 0 10px;
   display: flex;
@@ -377,56 +401,15 @@ export default {
 //   justify-content: space-between;
 // }
 
-.centerZone {
-  flex: 1 1 auto;
-  min-width: 140px;
+.playBtnContent {
   height: 100%;
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: center;
-  justify-content: center;
-  padding: 6px 0;
-  // 右侧功能按钮较宽，补偿偏移使控制键组接近视觉居中偏左
-  margin-right: 150px;
-}
-
-.transportRow {
   flex: none;
   display: flex;
   flex-flow: row nowrap;
   align-items: center;
   justify-content: center;
-  gap: 12px;
-}
-
-.progressRow {
-  flex: none;
-  width: 100%;
-  max-width: 500px;
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: center;
-  margin-top: 2px;
-}
-.time {
-  flex: none;
-  font-size: 11.5px;
-  font-variant-numeric: tabular-nums;
-  color: var(--color-550);
-  min-width: 36px;
-  &:first-child {
-    text-align: right;
-  }
-}
-.progress {
-  flex: auto;
-  min-width: 60px;
-  position: relative;
-  margin: 0 10px;
-  padding: 4px 0;
-  .progressBar {
-    height: 5px;
-  }
+  padding: 0 16px;
+  gap: 14px;
 }
 
 .playBtn {