+++ src/renderer/store/player/state.ts
@@ -69,5 +69,7 @@ export const playedList = window.lxData.playedList = shallowReactive<LX.Player.P
 
 export const tempPlayList = shallowReactive<LX.Player.PlayMusicInfo[]>([])
 
+export const isPlayQueueStopped = ref(false)
+
 window.lxData.playInfo = playInfo
 window.lxData.playMusicInfo = playMusicInfo